Display number of cards to learn as a badge in navigation Fixes #20
