+++ lib/src/decks/card_edit_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_flashcards/src/app.dart';
-import 'package:flutter_flashcards/src/base_layout.dart';
+import 'package:flutter_flashcards/src/layout/base_layout.dart';
 import 'package:flutter_flashcards/src/decks/card_edit.dart';
 import 'package:flutter_flashcards/src/widgets.dart';
 import '../model/cards.dart' as model;
